Issue 6 - renamed files.
