+++ src/sketchup-stl/exporter.rb
