+++ src/sketchup-stl/importer.rb
