+++ src/sketchup-stl/loader.rb
@@ -4,5 +4,5 @@
 
 require 'sketchup.rb'
 
-Sketchup::require 'sketchup-stl/jf_stl_importer'
-Sketchup::require 'sketchup-stl/skp_to_dxf'
+Sketchup::require 'sketchup-stl/importer'
+Sketchup::require 'sketchup-stl/exporter'
